# Reduced time taken to compute expected proposers